requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
mysql-connector-python==9.1.0
Werkzeug==3.1.3
python-dotenv==1.0.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
mysql-connector-python==9.1.0
Werkzeug==3.1.3
python-dotenv==1.0.1
